💄 select-room modal UI 스타일 업데이트
* 스크롤바 hidden
* 모달 헤더 버튼 추가
* 상단 헤더 고정
